refactoring for event objects to drive db records, not the other way around
